move BookBlock to own component
